Update patch set 3
Patch Set 3:

The build is broken after this commit:

In file included from ../../../src/dec/webp.c:15:

../../../src/dec/../webp/mux.h:49:24: error: webp/types.h: No such file or directory

This has been an issue for some time from the looks of things, but this is the first use of the header in this location. mux.h should #include ./types.h as the other installed headers do.

Patch-set: 3