Update patch set 1
Patch Set 1:

> Regarding your second comment, I am not sure I got your comment right. what should be the behavior with '--enable-decode-only' ? 1.) Generate 'libwebp_decode_only', instead of 'libwebp' library; or 2.) Generate both libwebp & libwebp_decode_only

> If it's option#1, I am fine with this suggestion. Option#2 is hard to implement as both libdsp & libutils (convenience libs) are impacted by this flag, unless we generate libdsp_decode_only & libutils_decode_only, in addition to libdsp & libutils.

Ah I thought (2) was the plan to begin with, i.e., split the enc/dec in dsp/utils then refer to only the decode libs for the new target and all of them for libwebp. Then we could have --disable-libwebp / --enable-libwebp-decode-only.

Really for now my biggest concern is producing a libwebp without all the normal symbols and having that distributed, that's why I prefer an opt-in and explicitly named lib.

If (2) is more work due to shared code modules, then (1) will probably be ok in the short term; may want to let skal chime in too...

Patch-set: 1