Update patch set 1
Patch Set 1:

Pascal -

High level Question. 
If we want to generate both libwebp & libwebpdecoder libs if --enable-decode-only is specified, the option name 'decode-only' is not apt.

What's the use case, when a user may require both the libs ?

Why not generate libwebp by default and libwebpdecoder when 'decode-only' option is specified ?

Patch-set: 1