Update patch set 2
Patch Set 2: (1 inline comment)

yes, we used libwebp-test-data, but we noticed that -noasm parameter is used for referent code and we did not know how to write some of our optimizations to be excluded when -noasm is present. do you have some suggestion how to do that? can we change C code and make some pointers to the functions? due to this problem we write our tests which used two executables, one for referent and one for test code and all images included from libwebp-test-data were passed.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0