Update patch set 1
Patch Set 1: No score

(4 inline comments)

The logic coded in the original patch is not equivalent to the one it replaced. You need to add 'if (streak > 3)' check before the cnts[] can be incremented.

If you add that logic, the compression density loss that was observed in the last patch is recovered. However, the compression speed is down 5-10%.

If the aim of this patch is to evaluate the logic equivalency between the C & MIPS32 code in patch (https://gerrit.chromium.org/gerrit/#/c/69595/), then the two logics are equivalent is this additional check 'if (streak > 3)' is added.

Patch-set: 1
Label: Code-Review=0