Update patch set 1
Patch Set 1: (9 inline comments)

can you hold on on this patch please?
I'm trying to see if we can work around the problem at a higher
level without having to introduce a new variant of the function.

Meanwhile, here are the comments i started entering, but it's just FYI.

Patch-set: 1