SmartRGBYUV: fix odd-width problem with pixel replication
rightmost pixel was missing a copy, which could lead to invalid read.

Also added a lower dimension of 4, below which we use the regular conversion.
This is to prevent corner cases, in addition to not being overkill.

(cherry picked from commit 2523aa73cb0aa337aaa5cd480a47a65476cdeb2c)

Change-Id: Iac12e7a3d74590f12fe8eeb1830b9891e61439f6