Update patch set 1
Patch Set 1:

> That's slippery and would affect all the code;

I was assuming we have a reference count of number of clang warnings before these changes (the false positive warnings). If so, we can increment the reference count in the automated gerrit tests.


> I haven't had a look at this, but nothing else has a problem with the pad fields, sounds like an Init() is missing...

The warning is coming from this line: "WebPAnimEncoderOptions enc_options = { { WHITE_COLOR, 0 }, 0, 0, 0, 0, 0 };"

I haven't really created an Init() method for this struct right now -- perhaps that's the way to go?

Patch-set: 1