Update patch set 1
Patch Set 1:

> I'd like to deal with this WEBP_EXTERN issue first, I think this is only used in the library currently, so doesn't need to be exposed like the others

As discussed offline with James, I tried removing the 'WEBP_EXTERN' from this method declaration, but it fails the building of 'webpmux' binary due to cross-dependency.

So, best thing we can do is to just push this to master branch (and not to the 0.5.1 branch).

I'll leave this open, so James can decide the best time to merge this patch (based on what makes branching 0.5.1 easier).

Patch-set: 1
Reviewer: Urvang Joshi <1123219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0