simplify HistogramCombineEntropyBin()
We only perform a single pass, and swap the final histograms
into the beginning of the array as we go. Therefore, they are
already at the correct place at the end of the pass.
-> HistogramCompactBins() is removed, we just truncate the array.

for low-effort (-m 0 -q 0):
The output is not bit-exact the same, because the histogram's
id are shuffled in a different order.

for non-low-effort: output is the same.

Change-Id: I9508c96dda0f8903c927a71b06af4e6490c3249c