[WIP] FastMBAnalyze() for quick i16/i4 decision
The decision is based on the DC4/DC16 comparison. The SSE
is computed in case of DC16 pred mode, and compared to all-DC4
modes for intra4.

It's only used for -m 0 and -m 1, where no full RD-opt is performed.
A tad slower, but gives much better results, close to -m 2.

Change-Id: I15f972db97ba4082cbd1dfd16bee3eb2eca701a8