smartYUV: improve initial state for faster convergence
For speed reason, the 'gray' plane was initialized with same
value for 2x2 block. But in some case (underlying camera noise, e.g.)
it could lead to instability, noise amplification during convergence,
and visible banding.
Using a precise (but slower) initialization solves the issue, and
since the convergence is faster, we might actually gain some speed.

Change-Id: I81c42101497e7096a8f60289d710f5a3bcb0ddea