Update patch set 38
Patch Set 38:

any timing? the code produced with clang looks reasonable (it often devolves to arm when dealing with lanes), but a quick test looks slower. The lower predictors result in very little arm code, so I don't know if we're getting much benefit using neon for 4 bytes (and we'll pay a transition cost between the 2 modes).

Patch-set: 38