Update patch set 1
Patch Set 1:

> Uploaded patch set 1.

For information, I am a GIMP developer and some contributors recently added support of WebP to GIMP. But after some tests on the kmin/kmax feature, and reading the doc (and comments of WebPAnimEncoderOptions), I didn't have the expected result to my test outputs.

I had to read the actual implementation to understand the real logics. My first idea was to propose only a comment update, but after some thoughts, I realized that the current special cases are completely messy and unlogical. The comment change would have to ask the developer to set kmax == 0 and any kmin > 0 to have all frames as keyframes.
There is also the fact that kmax == 1 is definitely a special case (since it cannot have any valid kmin with your rules) so it should be handled as such.

Moreover the proposed change makes much more sense as far as the English language is concerned (you could even guess what kmax == 0 and kmax == 1 would do).

So I do understand that here I am kind of proposing a behavior change of the API and this could be a problem. If that is really a no-go for this reason, then I could just revert to my original idea of just proposing a comment change.
But since WebP is still quite recent, I am hoping it is not too late to propose such a change which makes a lot more sense IMO.

Patch-set: 1
Label: Code-Review=0