Update patch set 3
Patch Set 3:

Hi Pascal!

 > Now, there's the problem of compatibility: 0.5.1 has been out for
 > quite some time already, and we would need to version this somehow.

I agree that may be a problem. This is why I was very unsure of what WebP developers would think of this proposed change. But I tried anyway because I really think current behavior is a mess. :-)

 > For GIMP, in particular, this would be mean special-casing the
 > calling
 > code depending on which libwebp version is installed on the system.
 
There won't be any problem nor any special-casing in GIMP in particular (other software, I don't know). First because no stable version ever got released with WebP so none expects any particular behavior with any set of parameters; second because we can just have a minimum requirement to the version of libWebP we want.

And last for the simple reason that anyway it just doesn't work well right now, IMO. When we set kmax == 1, then in our code kmin == 0, and it disables keyframing, whereas with any logics, it should mean the opposite (I was clearly expecting that any frame is a keyframe).
This is how I discovered the issue.

 > Adding Urvang here. Any opinion Urvang?
 > 
 > (and: maybe we could take this discussion to our mailing list,
 > webp-discuss@webmproject.org ?)

Patch-set: 3