Update patch set 3
Patch Set 3:

> Hi Pascal!
 > 
 > > Now, there's the problem of compatibility: 0.5.1 has been out for
 > > quite some time already, and we would need to version this
 > somehow.
 > 
 > I agree that may be a problem. This is why I was very unsure of
 > what WebP developers would think of this proposed change. But I
 > tried anyway because I really think current behavior is a mess. :-)
 > 
 > > For GIMP, in particular, this would be mean special-casing the
 > > calling
 > > code depending on which libwebp version is installed on the
 > system.
 > 
 > There won't be any problem nor any special-casing in GIMP in
 > particular (other software, I don't know). First because no stable
 > version ever got released with WebP so none expects any particular
 > behavior with any set of parameters; second because we can just
 > have a minimum requirement to the version of libWebP we want.
 > 
 > And last for the simple reason that anyway it just doesn't work
 > well right now, IMO. When we set kmax == 1, then in our code kmin
 > == 0, and it disables keyframing, whereas with any logics, it
 > should mean the opposite (I was clearly expecting that any frame is
 > a keyframe).
 > This is how I discovered the issue.
 > 
 > > Adding Urvang here. Any opinion Urvang?
 > >
 > > (and: maybe we could take this discussion to our mailing list,
 > > webp-discuss@webmproject.org ?)

I'd like to respond in two points:
(1) I think the source of the confusion is that it's not clear what "distance between keyframes" exactly means. I intended it to mean "the number of frames between two consecutive keyframes". So, if there was no frame between two consecutive keyframes, the distance would be 0.
(I can see why some of you may consider that to be a distance 1 instead. So I agree that the comment could be made clearer).

And this is why, 'kmax == 0' meant 'all frames are keyframes' (because at max 0 frames occur between any two consecutive keyframes).
[And, so I used the other special value 'kmin == 0' to signify the other special case].

(2) I'm definitely concerned about changing the behavior, that breaks compatibility with 0.5.1, as both of you already mentioned. So, I'd strongly suggest improving the comments.

Patch-set: 3