Update patch set 3
Patch Set 3:

> I'd like to respond in two points:
 > (1) I think the source of the confusion is that it's not clear what
 > "distance between keyframes" exactly means. I intended it to mean

This was exactly my first answer to Pascal when he contributed the patch to GIMP. I was not sure what was the distance between keyframes. But after checking a little libwebp code, I understood it as the expected "number of frames between two consecutive keyframes", i.e. your definition and yet…

 > "the number of frames between two consecutive keyframes". So, if
 > there was no frame between two consecutive keyframes, the distance
 > would be 0.
 > (I can see why some of you may consider that to be a distance 1
 > instead. So I agree that the comment could be made clearer).
 > 
 > And this is why, 'kmax == 0' meant 'all frames are keyframes'
 > (because at max 0 frames occur between any two consecutive
 > keyframes).

Hmmm… well ok if you really say so. But allow me to consider this far-fetched! :P
With the same "number of frames between 2 keyframes" definition, I understand kmax == 0 as: there are at most 0 frames between 2 keyframes, i.e. there are no keyframes (but the first). And kmax == 1 as there is at most 1 frame between 2 keyframes, i.e. every frame is a keyframe.

 > [And, so I used the other special value 'kmin == 0' to signify the
 > other special case].
 
As explained in my first message, the second problem with using both kmin and kmax as trigger values is: what happens when both kmin == kmax == 0?
Right now, "kmin" wins.
Worse, the test is actually kmin <= 0, which is not explained anywhere, so if you set kmax to 0 and kmin to just any lower value (because the doc says kmin must be lower than kmax, so you find this logical), you turn off insertion of key-frames (while expecting quite the opposite!). So you have to set for instance kmin == 1 and kmax == 0 to get your expected result, which is very ugly IMO and impose using some weird special casing in using code.

Of course, this could just be explained in the doc comment ("use kmax == 0 and kmin > 0 to make all frames key-frames"), but really… this is not proper, is it?

On the other hand, saying "kmax == 0 or 1 are 2 special values, in which cases kmin value is ignored" generates nicer code in using apps.

Of course, that may also be my warped idea of a nice code and you may disagree.

 > (2) I'm definitely concerned about changing the behavior, that
 > breaks compatibility with 0.5.1, as both of you already mentioned.
 > So, I'd strongly suggest improving the comments.

Is that your finale answer? Well I still argued for the sake of it. At least I will have tried and hopefully I could still change your mind. :-)

libWebP is still a recent library all-in-all. Is there already that much code based on this library that you can't update this with just a notice in the release note? Changing the behavior while the lib is still young may be the only real chance there is.

Patch-set: 3