Update patch set 5
Patch Set 5:

Ok so I added the commit email as an alternate email to the gmail account. In the CLA though, it still shows only the gmail email, but I apparently can't edit this, nor sign a new CLA (I get an error message saying I already signed one).
Even if I connect with the girinstud.io email, it only shows the gmail one.

So I assume that even if it shows only the gmail one, it's actually ok?

Patch-set: 5