Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Jehan,
> 
> now that this is officially modified, how are things looking like on the GIMP side? Do you need an update patch there?

I actually already have a stash somewhere with some minor update in GIMP UI relative to this change, so don't worry. But I didn't realize that libwebp 0.6 was released (I saw the news for the RCs, but I must have missed the announcement for the finale release). Anyway we kind of have an more-or-less-unsaid rule for dependency, which is that we can bump the versions in master if it is at least available in Debian Testing. But I see the libwebp in Debian Testing is 0.5.2. So I have to wait a little before I can bump our minimum dependency of libwebp anyway. :-)
Then I will push my small change.

Patch-set: 5