fix color smearing at very low quality setting
The aggressive bias was causing color bleeding and error propagation
on the U/V plane. The DC difference coded was forced to 0 too much,
causing errors to never be corrected.

We work around this by modulating the bias toward a more neutral
value a very low quality setting. This grows the file size a bit,
for a good reason.

Change-Id: I7fb895be6cb3813a7e514a515e8ff906b5b558a5